+++ src/test/compile-fail/unsendable-class.rs
@@ -16,11 +16,11 @@ use std::sync::mpsc::channel;
 use std::rc::Rc;
 
 struct foo {
-  i: int,
+  i: isize,
   j: Rc<String>,
 }
 
-fn foo(i:int, j: Rc<String>) -> foo {
+fn foo(i:isize, j: Rc<String>) -> foo {
     foo {
         i: i,
         j: j
@@ -31,6 +31,5 @@ fn main() {
   let cat = "kitty".to_string();
   let (tx, _) = channel();
   //~^ ERROR `core::marker::Send` is not implemented
-  //~^^ ERROR `core::marker::Send` is not implemented
   tx.send(foo(42, Rc::new(cat)));
 }
